+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Developer Changelog
+- Backend
+  - Added controller, model and routes for Budget Summary Report
+  - Changed report routes to use a common handler: getReport
+  - Changed project status most recent and budget summary report to match new standard
 
 ### November 16, 2022 (DESCW-727) Project Quarterly Billing Request Report
 - Backend